[#20] refactor mas-chat and mas-messaging
