--- config/api/gunicorn.conf.toml
@@ -1,5 +1,0 @@
-bind = "0.0.0.0:8000"
-workers = 8
-worker_class = "uvicorn.workers.UvicornWorker"
-accesslog = "-"
-errorlog = "-"
