--- config/common.toml
@@ -1,12 +1,0 @@
-[oauth2.google]
-client_id = "875870111546-ub0phvh6rcumm261nknla970lgpf5uve.apps.googleusercontent.com"
-project_id = "mas-server-382806"
-auth_uri = "https://accounts.google.com/o/oauth2/auth"
-token_uri = "https://oauth2.googleapis.com/token"
-auth_provider_x509_cert_url = "https://www.googleapis.com/oauth2/v1/certs"
-client_secret = "gAAAAABkMmQZNehPUGPL2hExWa5MxSRyvewu3QhyyD8quLFsvqbNz-2UC-rSZyiq_-qRfWz8mY8Y67rorP06sjAy9m9V7ZHiF-4BooJVLXsDAv5Nkhct1BoIPYPmRA1k47ERIVleJtRA"
-scopes=[
-    "https://www.googleapis.com/auth/userinfo.email",
-    "openid",
-    "https://www.googleapis.com/auth/userinfo.profile",
-]
