--- config/local.toml
@@ -1,18 +1,0 @@
-[database.mysql]
-username = "admin"
-password = "gAAAAABkKoPWU0bJwSREcQ01egg9091Xjbvs_TEaIg_7ZmxINB95kYRAKmdylqtaunYGkTFSa7HvV-G2vBk97m3fdbm63slTb1j0z-dypRthJarPbgzvhUg="
-endpoint = "mas-v2-database.cvemd7kvhgtq.ap-northeast-2.rds.amazonaws.com"
-port = 3306
-database_name = "mas"
-
-[oauth2.google]
-redirect_uri = "http://localhost:5500/login/"
-javascript_origins = [
-    "http://localhost:5500",
-]
-
-[kafka.producer]
-bootstrap.servers = "localhost"
-
-[kafka.consumer]
-bootstrap.servers = "localhost"
