--- config/websocket/gunicorn.conf.toml
@@ -1,5 +1,0 @@
-bind = "0.0.0.0:8001"
-workers = 1
-worker_class = "uvicorn.workers.UvicornWorker"
-accesslog = "-"
-errorlog = "-"
