+++ mas-api/api/user/repository/user_repository.py
@@ -1,9 +1,8 @@
-from sqlalchemy import delete, select, update
-
 from api.database.database_connection_manager import DatabaseConnectionManager
 from api.user.entity.user import User
 from api.user.exception.user_exception import UserNotFoundException
 from api.utils.datetime_utils import get_now_datetime_by_timezone
+from sqlalchemy import delete, select, update
 
 
 class UserRepository:
