+++ mas-chat/chat/dto/script_dto.py
@@ -1,0 +1,9 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class ScriptDto(BaseModel):
+    name: str
+    content: str
+    created_at: datetime
