+++ mas-chat/chat/mas_chat_application.py
@@ -1,0 +1,6 @@
+from chat.utils.controller_utils import create_app
+from chat.utils.gunicorn_utils import MasWebsocketChatServer
+
+app = create_app(".", False)
+
+MasWebsocketChatServer(app, "./config/gunicorn.conf.toml").run()
