--- mas-chat/chat/mas_websocket_application.py
@@ -1,6 +1,0 @@
-from mas.utils.controller_utils import create_app
-from mas.utils.gunicorn_utils import MASMultiProcessingServer
-
-app = create_app("./api/websocket", False)
-
-MASMultiProcessingServer(app, "./config/websocket/gunicorn.conf.toml").run()
