--- mas-chat/chat/script/dto/script_response.py
@@ -1,9 +1,0 @@
-from datetime import datetime
-
-from pydantic import BaseModel
-
-
-class ScriptResponse(BaseModel):
-    name: str
-    content: str
-    created_at: datetime
