+++ mas-chat/chat/utils/const.py
@@ -1,0 +1,15 @@
+"""
+Constant Variables
+"""
+
+LOGGERS = [
+    "uvicorn",
+    "uvicorn.asgi",
+    "uvicorn.access",
+    "uvicorn.error",
+    "gunicorn",
+    "gunicorn.access",
+    "gunicorn.error",
+    "gunicorn.http.wsgi",
+    "gunicorn.http",
+]
