+++ mas-messaging/messaging/database/mysql_connection_manager.py
@@ -3,6 +3,8 @@ from contextlib import AbstractContextManager, asynccontextmanager
 from logging import getLogger
 from typing import Callable
 
+from messaging.database.database_connection_manager import DatabaseConnectionManager
+from messaging.utils.config import Config
 from sqlalchemy import NullPool
 from sqlalchemy.ext.asyncio import (
     AsyncSession,
@@ -10,9 +12,6 @@ from sqlalchemy.ext.asyncio import (
     async_sessionmaker,
     create_async_engine,
 )
-
-from api.database.database_connection_manager import DatabaseConnectionManager
-from api.utils.config import Config
 
 logger = getLogger()
