+++ mas-messaging/messaging/utils/config.py
@@ -18,7 +18,7 @@ class Config:
     """
 
     database = None
-    oauth2 = None
+    kafka = None
 
     def __init__(self, phase: str):
         with open("./config/local.toml", "rb") as f:
