+++ mas-messaging/messaging/utils/const.py
@@ -1,0 +1,1 @@
+DATABASE_DRIVER_CLASS = {"mysql": "mysql+aiomysql"}
